# api-documentation.yaml
openapi: 3.0.3
info:
  title: Organization Service API
  description: |
    API documentation for the Organization Service which manages organizations, patients, and users.
    
    ## Authentication
    All endpoints (except /health) require Bearer token authentication.
    Header: `Authorization: Bearer <your-token>`
    
    ## For SUPER_ADMIN
    Add `X-Organization-ID` header to specify target organization.
    
    ## For ORG_ADMIN
    Uses organization from token automatically. Do NOT send X-Organization-ID header.
  version: 1.0.0
  contact:
    name: WailSalutem Health Care
    email: support@wailsalutem.com
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.wailsalutem.com
    description: Production
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        schema_name:
          type: string
        contact_email:
          type: string
          format: email
        contact_phone:
          type: string
        address:
          type: string
        status:
          type: string
        settings:
          type: object
        created_at:
          type: string
          format: date-time
    CreateOrganizationRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: Healthcare Center A
        contact_email:
          type: string
          format: email
          example: contact@healthcare.com
        contact_phone:
          type: string
          example: "+1234567890"
        address:
          type: string
          example: 123 Main St, City, State 12345
        settings:
          type: object
    UpdateOrganizationRequest:
      type: object
      properties:
        name:
          type: string
        contact_email:
          type: string
          format: email
        contact_phone:
          type: string
        address:
          type: string
...
paths:
...
  /organizations:
    post:
      tags:
        - Organizations
      summary: Create organization
      description: Create a new organization (SUPER_ADMIN only)
      operationId: createOrganization
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationRequest'
      responses:
        '201':
          description: Organization created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

    get:
      tags:
        - Organizations
      summary: List organizations
      description: |
        List organizations based on role:
        - SUPER_ADMIN: Returns all organizations
        - ORG_ADMIN: Returns only their own organization
      operationId: listOrganizations
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of organizations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Organization'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /organizations/{id}:
    get:
      tags:
        - Organizations
      summary: Get organization by ID
      description: |
        - SUPER_ADMIN: Can view any organization
        - ORG_ADMIN: Can only view their own organization
      operationId: getOrganization
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Organization details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found

    put:
      tags:
        - Organizations
      summary: Update organization (full)
      description: SUPER_ADMIN only. ORG_ADMIN gets 403 Forbidden.
      operationId: updateOrganizationPut
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrganizationRequest'
      responses:
        '200':
          description: Organization updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found

    patch:
      tags:
        - Organizations
      summary: Update organization (partial)
      description: SUPER_ADMIN only. ORG_ADMIN gets 403 Forbidden.
      operationId: updateOrganizationPatch
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrganizationRequest'
      responses:
        '200':
          description: Organization updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found

    delete:
      tags:
        - Organizations
      summary: Delete organization
      description: SUPER_ADMIN only
      operationId: deleteOrganization
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Organization deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
...